Lift URL util to avoid x-container calls
