+++ ui/apps/platform/src/Containers/NetworkGraph/utils/networkGraphURLUtils.ts
@@ -1,27 +1,4 @@
-import { networkBasePath } from 'routePaths';
-import { getQueryString } from 'utils/queryStringUtils';
 import type { SearchFilter } from 'types/search';
-
-type GetURLLinkToDeploymentParams = {
-    cluster: string;
-    namespace: string;
-    deploymentId: string;
-};
-
-export function getURLLinkToDeployment({
-    cluster,
-    namespace,
-    deploymentId,
-}: GetURLLinkToDeploymentParams) {
-    const queryString = getQueryString({
-        s: {
-            Cluster: cluster,
-            Namespace: namespace,
-        },
-    });
-    const networkGraphLink = `${networkBasePath}/deployment/${deploymentId}${queryString}`;
-    return networkGraphLink;
-}
 
 export function getPropertiesForAnalytics(searchFilter: SearchFilter) {
     const cluster = searchFilter?.Cluster?.toString() ? 1 : 0;
